requirements.txt
python >= 3.6.13
pytorch >= 1.8.1
pandas >= 1.1.5
numpy >= 1.19.5
dgl >= 0.6.1
pyyaml >= 5.4.1
tqdm >= 4.61.0
pybind11 >= 2.6.2
g++ >= 7.5.0
openmp >= 201511